LEARN BOT/swfm/requirements.txt
# python-telegram-bot
python-dotenv
mysql-connector-python
pytelegrambotapi
panda
logging
telegram
watchdog
# os